requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncer==0.0.11
attrs==25.3.0
beautifulsoup4==4.14.3
brotli==1.2.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.2
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
easydict==1.13
et_xmlfile==2.0.0
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.8.0
fsspec==2025.12.0
gdown==5.2.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gradio==6.0.2
gradio_client==2.0.1
greenlet==3.2.3
groovy==0.1.2
gspread==6.2.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.3
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
orjson==3.11.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
pooch==1.8.2
propcache==0.4.1
protobuf==6.31.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMatting==1.1.14
PyMsgBox==1.0.9
pynput==1.8.1
pyobjc-core==12.0
pyobjc-framework-ApplicationServices==11.1
pyobjc-framework-Cocoa==12.0
pyobjc-framework-CoreText==11.1
pyobjc-framework-Quartz==12.0
pyparsing==3.2.3
pyperclip==1.9.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.26.0
rsa==4.9.1
rubicon-objc==0.5.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.0
selenium==4.38.0
semantic-version==2.10.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.50.0
stringzilla==4.4.0
sympy==1.14.0
tifffile==2025.6.11
timm==1.0.22
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.57.3
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.22.0
